requirements.txt
fastapi==0.118.0
uvicorn[standard]==0.36.0
python-multipart==0.0.20
langchain==0.3.12
langchain-google-genai==2.0.8
langchain-community==0.3.12
chromadb==1.1.0
google-generativeai==0.8.3
python-dotenv==1.0.0
pypdf==6.1.1
python-docx==1.2.0
pydantic==2.10.4
pydantic-settings==2.7.1
sentence-transformers==3.0.1